learning rate decay impl

+++ oneflow/core/kernel/momentum_model_update_kernel.cpp
@@ -11,7 +11,7 @@ void MomentumMdUpdateKernel<device_type, T>::UpdateModel(
   Blob* momentum_blob = BnInOp2Blob("momentum");
   const NormalModelUpdateOpUserConf& conf =
       this->op_conf().normal_mdupdt_conf().user_conf();
-  float learning_rate = conf.learning_rate();
+  double learning_rate = conf.learning_rate();
   if (conf.has_lr_decay()) {
     learning_rate = GetDecayedLearningRate(conf.lr_decay(), learning_rate,
                                            next_model_vid - 1);
